Move MQTT to ManualActivity
