--- android/app/src/main/java/com/example/android/MqttBackgroundProcess.java
@@ -1,128 +1,0 @@
-package com.example.android;
-
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
-import android.os.Bundle;
-import android.util.Log;
-import android.widget.Button;
-import android.widget.Toast;
-import android.app.Application;
-import android.content.Context;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import org.eclipse.paho.client.mqttv3.IMqttActionListener;
-import org.eclipse.paho.client.mqttv3.IMqttDeliveryToken;
-import org.eclipse.paho.client.mqttv3.IMqttToken;
-import org.eclipse.paho.client.mqttv3.MqttCallback;
-import org.eclipse.paho.client.mqttv3.MqttMessage;
-
-public class MqttBackgroundProcess extends BroadcastReceiver {
-    private static final String TAG = "SmartcarMqttController";
-    private static final String EXTERNAL_MQTT_BROKER = "aerostun.dev";
-    private static final String LOCALHOST = "10.0.2.2";
-    private static final String MQTT_SERVER = "tcp://" + EXTERNAL_MQTT_BROKER + ":1883";
-    private static final String THROTTLE_CONTROL = "/smartcar/control/throttle";
-    private static final String STEERING_CONTROL = "/smartcar/control/steering";
-
-    private static final int MOVEMENT_SPEED = 70;
-    private static final int IDLE_SPEED = 0;
-    private static final int STRAIGHT_ANGLE = 0;
-    private static final int STEERING_ANGLE = 50;
-    private static final int QOS = 1;
-
-    @Override
-    public void onReceive(Context context, Intent intent) {
-        mMqttClient = new MqttClient(context, MQTT_SERVER, TAG);
-        connectToMqttBroker();
-
-    }
-
-    private MqttClient mMqttClient;
-    private boolean isConnected = false;
-
-
-    private void connectToMqttBroker() {
-        if (!isConnected) {
-            mMqttClient.connect(TAG, "", new IMqttActionListener() {
-                @Override
-                public void onSuccess(IMqttToken asyncActionToken) {
-                    isConnected = true;
-
-                    final String successfulConnection = "Connected to MQTT broker";
-                    Log.i(TAG, successfulConnection);
-                    Toast.makeText(getApplicationContext(), successfulConnection, Toast.LENGTH_SHORT).show();
-
-                    mMqttClient.subscribe("/smartcar/ultrasound/front", QOS, null);
-
-                }
-
-                @Override
-                public void onFailure(IMqttToken asyncActionToken, Throwable exception) {
-                    final String failedConnection = "Failed to connect to MQTT broker";
-                    Log.e(TAG, failedConnection);
-                    Toast.makeText(getApplicationContext(), failedConnection, Toast.LENGTH_SHORT).show();
-                }
-            }, new MqttCallback() {
-                @Override
-                public void connectionLost(Throwable cause) {
-                    isConnected = false;
-
-                    final String connectionLost = "Connection to MQTT broker lost";
-                    Log.w(TAG, connectionLost);
-                    Toast.makeText(getApplicationContext(), connectionLost, Toast.LENGTH_SHORT).show();
-                }
-
-                @Override
-                public void messageArrived(String topic, MqttMessage message) throws Exception {
-
-                    Log.i(TAG, "[MQTT] Topic: " + topic + " | Message: " + message.toString());
-                }
-
-                @Override
-                public void deliveryComplete(IMqttDeliveryToken token) {
-                    Log.d(TAG, "Message delivered");
-                }
-            });
-        }
-
-    }
-    void drive(int throttleSpeed, int steeringAngle, String actionDescription) {
-        if (!isConnected) {
-            final String notConnected = "Not connected (yet)";
-            Log.e(TAG, notConnected);
-            Toast.makeText(getApplicationContext(), notConnected, Toast.LENGTH_SHORT).show();
-            return;
-        }
-        Log.i(TAG, actionDescription);
-        mMqttClient.publish(THROTTLE_CONTROL, Integer.toString(throttleSpeed), QOS, null);
-        mMqttClient.publish(STEERING_CONTROL, Integer.toString(steeringAngle), QOS, null);
-    }
-
-    private Context getApplicationContext() {
-        Context appContext = getApplicationContext();
-        return appContext;
-    }
-
-    public void forward() {
-        drive(MOVEMENT_SPEED, STRAIGHT_ANGLE, "Moving forward");
-    }
-
-    public void left() {
-        drive(MOVEMENT_SPEED, -STEERING_ANGLE, "Moving forward left");
-    }
-
-    public void stop() {
-        drive(IDLE_SPEED, STRAIGHT_ANGLE, "Stopping");
-    }
-
-    public void right() {
-        drive(MOVEMENT_SPEED, STEERING_ANGLE, "Moving forward left");
-    }
-
-    public void backward() {
-        drive(-MOVEMENT_SPEED, STRAIGHT_ANGLE, "Moving backward");
-    }
-
-}
